Move ccd relaited objects to own package
