+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/Authenticator.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.bulkscan.orchestrator.services;
+package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd;
 
 import uk.gov.hmcts.reform.idam.client.models.UserDetails;
